Add otponly mode, where only the OTP password is required
When invoked as login_otp_only, login_totp_only, or login_hotp_only,
the user's system password is not required to be appended to the
TOTP password.

This is useful with OpenSSH's AuthenticationMethods option which can
require, for example, a user's SSH key first, then require
keyboard-interactive to call login_otp_only to require an OTP from
the user.